+++ app/src/content.config.ts
@@ -8,7 +8,7 @@
  * SPDX-License-Identifier: UNLICENSED
  */
 import { join } from "node:path";
-import { invariant } from "@ariakit/core/utils/misc";
+import { invariant } from "@ariakit/utils";
 import { glob } from "astro/loaders";
 import { z } from "astro/zod";
 import { defineCollection, reference } from "astro:content";
@@ -102,13 +102,19 @@ const references = defineCollection({
   loader: jsdoc(
     {
       framework: "react",
-      corePath: join(import.meta.dirname, "../../packages/ariakit-react-core"),
+      corePath: join(
+        import.meta.dirname,
+        "../../packages/ariakit-react-components",
+      ),
       packagePath: join(import.meta.dirname, "../../packages/ariakit-react"),
       // watch: true,
     },
     {
       framework: "solid",
-      corePath: join(import.meta.dirname, "../../packages/ariakit-solid-core"),
+      corePath: join(
+        import.meta.dirname,
+        "../../packages/ariakit-solid-components",
+      ),
       packagePath: join(import.meta.dirname, "../../packages/ariakit-solid"),
       // watch: true,
     },