+++ examples/composite-virtual-focus-events/test.ts
@@ -13,7 +13,7 @@ test("events", async () => {
   externalButton.textContent = "External button";
   document.body.append(externalButton);
 
-  const log = vi.spyOn(console, "log").mockImplementation(() => {});
+  using log = vi.spyOn(console, "log").mockImplementation(() => {});
 
   await press.Tab();
   await press.Tab();
@@ -53,8 +53,6 @@ test("events", async () => {
       "event: focus | currentTarget: toolbar | target: item-2 | relatedTarget: toolbar",
     ]
   `);
-
-  log.mockReset();
 
   externalButton.remove();
 });