+++ packages/ariakit-react/src/combobox.ts
@@ -1,76 +1,76 @@
 export type {
   ComboboxOptions,
   ComboboxProps,
-} from "@ariakit/react-core/combobox/combobox";
-export { Combobox } from "@ariakit/react-core/combobox/combobox";
+} from "@ariakit/react-components/combobox/combobox";
+export { Combobox } from "@ariakit/react-components/combobox/combobox";
 export type {
   ComboboxCancelOptions,
   ComboboxCancelProps,
-} from "@ariakit/react-core/combobox/combobox-cancel";
-export { ComboboxCancel } from "@ariakit/react-core/combobox/combobox-cancel";
-export { useComboboxContext } from "@ariakit/react-core/combobox/combobox-context";
+} from "@ariakit/react-components/combobox/combobox-cancel";
+export { ComboboxCancel } from "@ariakit/react-components/combobox/combobox-cancel";
+export { useComboboxContext } from "@ariakit/react-components/combobox/combobox-context";
 export type {
   ComboboxDisclosureOptions,
   ComboboxDisclosureProps,
-} from "@ariakit/react-core/combobox/combobox-disclosure";
-export { ComboboxDisclosure } from "@ariakit/react-core/combobox/combobox-disclosure";
+} from "@ariakit/react-components/combobox/combobox-disclosure";
+export { ComboboxDisclosure } from "@ariakit/react-components/combobox/combobox-disclosure";
 export type {
   ComboboxGroupOptions,
   ComboboxGroupProps,
-} from "@ariakit/react-core/combobox/combobox-group";
-export { ComboboxGroup } from "@ariakit/react-core/combobox/combobox-group";
+} from "@ariakit/react-components/combobox/combobox-group";
+export { ComboboxGroup } from "@ariakit/react-components/combobox/combobox-group";
 export type {
   ComboboxGroupLabelOptions,
   ComboboxGroupLabelProps,
-} from "@ariakit/react-core/combobox/combobox-group-label";
-export { ComboboxGroupLabel } from "@ariakit/react-core/combobox/combobox-group-label";
+} from "@ariakit/react-components/combobox/combobox-group-label";
+export { ComboboxGroupLabel } from "@ariakit/react-components/combobox/combobox-group-label";
 export type {
   ComboboxItemOptions,
   ComboboxItemProps,
-} from "@ariakit/react-core/combobox/combobox-item";
-export { ComboboxItem } from "@ariakit/react-core/combobox/combobox-item";
+} from "@ariakit/react-components/combobox/combobox-item";
+export { ComboboxItem } from "@ariakit/react-components/combobox/combobox-item";
 export type {
   ComboboxItemCheckOptions,
   ComboboxItemCheckProps,
-} from "@ariakit/react-core/combobox/combobox-item-check";
-export { ComboboxItemCheck } from "@ariakit/react-core/combobox/combobox-item-check";
+} from "@ariakit/react-components/combobox/combobox-item-check";
+export { ComboboxItemCheck } from "@ariakit/react-components/combobox/combobox-item-check";
 export type {
   ComboboxItemValueOptions,
   ComboboxItemValueProps,
-} from "@ariakit/react-core/combobox/combobox-item-value";
-export { ComboboxItemValue } from "@ariakit/react-core/combobox/combobox-item-value";
+} from "@ariakit/react-components/combobox/combobox-item-value";
+export { ComboboxItemValue } from "@ariakit/react-components/combobox/combobox-item-value";
 export type {
   ComboboxLabelOptions,
   ComboboxLabelProps,
-} from "@ariakit/react-core/combobox/combobox-label";
-export { ComboboxLabel } from "@ariakit/react-core/combobox/combobox-label";
+} from "@ariakit/react-components/combobox/combobox-label";
+export { ComboboxLabel } from "@ariakit/react-components/combobox/combobox-label";
 export type {
   ComboboxListOptions,
   ComboboxListProps,
-} from "@ariakit/react-core/combobox/combobox-list";
-export { ComboboxList } from "@ariakit/react-core/combobox/combobox-list";
+} from "@ariakit/react-components/combobox/combobox-list";
+export { ComboboxList } from "@ariakit/react-components/combobox/combobox-list";
 export type {
   ComboboxPopoverOptions,
   ComboboxPopoverProps,
-} from "@ariakit/react-core/combobox/combobox-popover";
-export { ComboboxPopover } from "@ariakit/react-core/combobox/combobox-popover";
-export type { ComboboxProviderProps } from "@ariakit/react-core/combobox/combobox-provider";
-export { ComboboxProvider } from "@ariakit/react-core/combobox/combobox-provider";
+} from "@ariakit/react-components/combobox/combobox-popover";
+export { ComboboxPopover } from "@ariakit/react-components/combobox/combobox-popover";
+export type { ComboboxProviderProps } from "@ariakit/react-components/combobox/combobox-provider";
+export { ComboboxProvider } from "@ariakit/react-components/combobox/combobox-provider";
 export type {
   ComboboxRowOptions,
   ComboboxRowProps,
-} from "@ariakit/react-core/combobox/combobox-row";
-export { ComboboxRow } from "@ariakit/react-core/combobox/combobox-row";
+} from "@ariakit/react-components/combobox/combobox-row";
+export { ComboboxRow } from "@ariakit/react-components/combobox/combobox-row";
 export type {
   ComboboxSeparatorOptions,
   ComboboxSeparatorProps,
-} from "@ariakit/react-core/combobox/combobox-separator";
-export { ComboboxSeparator } from "@ariakit/react-core/combobox/combobox-separator";
+} from "@ariakit/react-components/combobox/combobox-separator";
+export { ComboboxSeparator } from "@ariakit/react-components/combobox/combobox-separator";
 export type {
   ComboboxStore,
   ComboboxStoreProps,
   ComboboxStoreState,
-} from "@ariakit/react-core/combobox/combobox-store";
-export { useComboboxStore } from "@ariakit/react-core/combobox/combobox-store";
-export type { ComboboxValueProps } from "@ariakit/react-core/combobox/combobox-value";
-export { ComboboxValue } from "@ariakit/react-core/combobox/combobox-value";
+} from "@ariakit/react-components/combobox/combobox-store";
+export { useComboboxStore } from "@ariakit/react-components/combobox/combobox-store";
+export type { ComboboxValueProps } from "@ariakit/react-components/combobox/combobox-value";
+export { ComboboxValue } from "@ariakit/react-components/combobox/combobox-value";