+++ packages/ariakit-react/src/command.ts
@@ -1,5 +1,5 @@
 export type {
   CommandOptions,
   CommandProps,
-} from "@ariakit/react-core/command/command";
-export { Command } from "@ariakit/react-core/command/command";
+} from "@ariakit/react-components/command/command";
+export { Command } from "@ariakit/react-components/command/command";