+++ packages/ariakit-react/src/composite.ts
@@ -1,49 +1,49 @@
 export type {
   CompositeOptions,
   CompositeProps,
-} from "@ariakit/react-core/composite/composite";
-export { Composite } from "@ariakit/react-core/composite/composite";
-export { useCompositeContext } from "@ariakit/react-core/composite/composite-context";
+} from "@ariakit/react-components/composite/composite";
+export { Composite } from "@ariakit/react-components/composite/composite";
+export { useCompositeContext } from "@ariakit/react-components/composite/composite-context";
 export type {
   CompositeGroupOptions,
   CompositeGroupProps,
-} from "@ariakit/react-core/composite/composite-group";
-export { CompositeGroup } from "@ariakit/react-core/composite/composite-group";
+} from "@ariakit/react-components/composite/composite-group";
+export { CompositeGroup } from "@ariakit/react-components/composite/composite-group";
 export type {
   CompositeGroupLabelOptions,
   CompositeGroupLabelProps,
-} from "@ariakit/react-core/composite/composite-group-label";
-export { CompositeGroupLabel } from "@ariakit/react-core/composite/composite-group-label";
+} from "@ariakit/react-components/composite/composite-group-label";
+export { CompositeGroupLabel } from "@ariakit/react-components/composite/composite-group-label";
 export type {
   CompositeHoverOptions,
   CompositeHoverProps,
-} from "@ariakit/react-core/composite/composite-hover";
-export { CompositeHover } from "@ariakit/react-core/composite/composite-hover";
+} from "@ariakit/react-components/composite/composite-hover";
+export { CompositeHover } from "@ariakit/react-components/composite/composite-hover";
 export type {
   CompositeItemOptions,
   CompositeItemProps,
-} from "@ariakit/react-core/composite/composite-item";
-export { CompositeItem } from "@ariakit/react-core/composite/composite-item";
-export type { CompositeProviderProps } from "@ariakit/react-core/composite/composite-provider";
-export { CompositeProvider } from "@ariakit/react-core/composite/composite-provider";
+} from "@ariakit/react-components/composite/composite-item";
+export { CompositeItem } from "@ariakit/react-components/composite/composite-item";
+export type { CompositeProviderProps } from "@ariakit/react-components/composite/composite-provider";
+export { CompositeProvider } from "@ariakit/react-components/composite/composite-provider";
 export type {
   CompositeRowOptions,
   CompositeRowProps,
-} from "@ariakit/react-core/composite/composite-row";
-export { CompositeRow } from "@ariakit/react-core/composite/composite-row";
+} from "@ariakit/react-components/composite/composite-row";
+export { CompositeRow } from "@ariakit/react-components/composite/composite-row";
 export type {
   CompositeSeparatorOptions,
   CompositeSeparatorProps,
-} from "@ariakit/react-core/composite/composite-separator";
-export { CompositeSeparator } from "@ariakit/react-core/composite/composite-separator";
+} from "@ariakit/react-components/composite/composite-separator";
+export { CompositeSeparator } from "@ariakit/react-components/composite/composite-separator";
 export type {
   CompositeStore,
   CompositeStoreProps,
   CompositeStoreState,
-} from "@ariakit/react-core/composite/composite-store";
-export { useCompositeStore } from "@ariakit/react-core/composite/composite-store";
+} from "@ariakit/react-components/composite/composite-store";
+export { useCompositeStore } from "@ariakit/react-components/composite/composite-store";
 export type {
   CompositeTypeaheadOptions,
   CompositeTypeaheadProps,
-} from "@ariakit/react-core/composite/composite-typeahead";
-export { CompositeTypeahead } from "@ariakit/react-core/composite/composite-typeahead";
+} from "@ariakit/react-components/composite/composite-typeahead";
+export { CompositeTypeahead } from "@ariakit/react-components/composite/composite-typeahead";