+++ packages/ariakit-react/src/hovercard.ts
@@ -1,44 +1,44 @@
 export type {
   HovercardOptions,
   HovercardProps,
-} from "@ariakit/react-core/hovercard/hovercard";
-export { Hovercard } from "@ariakit/react-core/hovercard/hovercard";
+} from "@ariakit/react-components/hovercard/hovercard";
+export { Hovercard } from "@ariakit/react-components/hovercard/hovercard";
 export type {
   HovercardAnchorOptions,
   HovercardAnchorProps,
-} from "@ariakit/react-core/hovercard/hovercard-anchor";
-export { HovercardAnchor } from "@ariakit/react-core/hovercard/hovercard-anchor";
+} from "@ariakit/react-components/hovercard/hovercard-anchor";
+export { HovercardAnchor } from "@ariakit/react-components/hovercard/hovercard-anchor";
 export type {
   HovercardArrowOptions,
   HovercardArrowProps,
-} from "@ariakit/react-core/hovercard/hovercard-arrow";
-export { HovercardArrow } from "@ariakit/react-core/hovercard/hovercard-arrow";
-export { useHovercardContext } from "@ariakit/react-core/hovercard/hovercard-context";
+} from "@ariakit/react-components/hovercard/hovercard-arrow";
+export { HovercardArrow } from "@ariakit/react-components/hovercard/hovercard-arrow";
+export { useHovercardContext } from "@ariakit/react-components/hovercard/hovercard-context";
 export type {
   HovercardDescriptionOptions,
   HovercardDescriptionProps,
-} from "@ariakit/react-core/hovercard/hovercard-description";
-export { HovercardDescription } from "@ariakit/react-core/hovercard/hovercard-description";
+} from "@ariakit/react-components/hovercard/hovercard-description";
+export { HovercardDescription } from "@ariakit/react-components/hovercard/hovercard-description";
 export type {
   HovercardDisclosureOptions,
   HovercardDisclosureProps,
-} from "@ariakit/react-core/hovercard/hovercard-disclosure";
-export { HovercardDisclosure } from "@ariakit/react-core/hovercard/hovercard-disclosure";
+} from "@ariakit/react-components/hovercard/hovercard-disclosure";
+export { HovercardDisclosure } from "@ariakit/react-components/hovercard/hovercard-disclosure";
 export type {
   HovercardDismissOptions,
   HovercardDismissProps,
-} from "@ariakit/react-core/hovercard/hovercard-dismiss";
-export { HovercardDismiss } from "@ariakit/react-core/hovercard/hovercard-dismiss";
+} from "@ariakit/react-components/hovercard/hovercard-dismiss";
+export { HovercardDismiss } from "@ariakit/react-components/hovercard/hovercard-dismiss";
 export type {
   HovercardHeadingOptions,
   HovercardHeadingProps,
-} from "@ariakit/react-core/hovercard/hovercard-heading";
-export { HovercardHeading } from "@ariakit/react-core/hovercard/hovercard-heading";
-export type { HovercardProviderProps } from "@ariakit/react-core/hovercard/hovercard-provider";
-export { HovercardProvider } from "@ariakit/react-core/hovercard/hovercard-provider";
+} from "@ariakit/react-components/hovercard/hovercard-heading";
+export { HovercardHeading } from "@ariakit/react-components/hovercard/hovercard-heading";
+export type { HovercardProviderProps } from "@ariakit/react-components/hovercard/hovercard-provider";
+export { HovercardProvider } from "@ariakit/react-components/hovercard/hovercard-provider";
 export type {
   HovercardStore,
   HovercardStoreProps,
   HovercardStoreState,
-} from "@ariakit/react-core/hovercard/hovercard-store";
-export { useHovercardStore } from "@ariakit/react-core/hovercard/hovercard-store";
+} from "@ariakit/react-components/hovercard/hovercard-store";
+export { useHovercardStore } from "@ariakit/react-components/hovercard/hovercard-store";