+++ packages/ariakit-react/src/separator.ts
@@ -1,5 +1,5 @@
 export type {
   SeparatorOptions,
   SeparatorProps,
-} from "@ariakit/react-core/separator/separator";
-export { Separator } from "@ariakit/react-core/separator/separator";
+} from "@ariakit/react-components/separator/separator";
+export { Separator } from "@ariakit/react-components/separator/separator";