+++ packages/ariakit-solid/src/role.ts
@@ -1,2 +1,5 @@
-export type { RoleOptions, RoleProps } from "@ariakit/solid-core/role/role";
-export { Role } from "@ariakit/solid-core/role/role";
+export type {
+  RoleOptions,
+  RoleProps,
+} from "@ariakit/solid-components/role/role";
+export { Role } from "@ariakit/solid-components/role/role";