+++ packages/ariakit-test/src/mouse-up.ts
@@ -1,5 +1,4 @@
-import { isVisible } from "@ariakit/core/utils/dom";
-import { invariant } from "@ariakit/core/utils/misc";
+import { isVisible, invariant } from "@ariakit/utils";
 import { wrapAsync } from "./__utils.ts";
 import { dispatch } from "./dispatch.ts";