+++ CHANGELOG.md
@@ -2,10 +2,16 @@
 
 Change logs are organized by package:
 
-- [@ariakit/core](packages/ariakit-core/CHANGELOG.md)
+- [@ariakit/components](packages/ariakit-components/CHANGELOG.md)
 - [@ariakit/react](packages/ariakit-react/CHANGELOG.md)
-- [@ariakit/react-core](packages/ariakit-react-core/CHANGELOG.md)
+- [@ariakit/react-components](packages/ariakit-react-components/CHANGELOG.md)
+- [@ariakit/react-store](packages/ariakit-react-store/CHANGELOG.md)
+- [@ariakit/react-utils](packages/ariakit-react-utils/CHANGELOG.md)
 - [@ariakit/solid](packages/ariakit-solid/CHANGELOG.md)
-- [@ariakit/solid-core](packages/ariakit-solid-core/CHANGELOG.md)
+- [@ariakit/solid-components](packages/ariakit-solid-components/CHANGELOG.md)
+- [@ariakit/solid-store](packages/ariakit-solid-store/CHANGELOG.md)
+- [@ariakit/solid-utils](packages/ariakit-solid-utils/CHANGELOG.md)
+- [@ariakit/store](packages/ariakit-store/CHANGELOG.md)
 - [@ariakit/tailwind](packages/ariakit-tailwind/CHANGELOG.md)
 - [@ariakit/test](packages/ariakit-test/CHANGELOG.md)
+- [@ariakit/utils](packages/ariakit-utils/CHANGELOG.md)